requirements.txt
pigpio
aiohttp
numpy
scikit-learn
pyserial